requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.0
attrs==20.3.0
autopep8==1.5.4
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dateparser==1.0.0
Django==3.0.5
djongo==1.3.3
dnspython==2.0.0
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
python-bsonjs==0.2.0
python-dateutil==2.8.1
pytz==2020.4
regex==2020.10.28
requests==2.24.0
six==1.15.0
sqlparse==0.2.4
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
whitenoise==5.2.0
